Fix permissions
This commit makes a couple of fixes related to permissions:

1. Removes the permissions controlling usage of the permissions check in
the Inventory UI. This plugin doesn't need its use to be restricted, if
a user can access the inventory area, the should be authorised to look
up permissions.
2. Fix the Settings permission name. The permission name should reflect
the full name of the module.
3. Remove the permission check around the display of the Inventory
Action button link. As discussed in this thread:
https://folio-project.slack.com/archives/C210UCHQ9/p1618836631336300 the
plugin itself determines whether the link is rendered or not by calling
the passed renderTrigger prop. So we don't need to wrap the <Button>
component in another permissions check.