onSubmit -> handleSubmit
The change in 657c1a9bc53a716ee6b35f9efbd65d665b101095 broke settings
savings. Despite rooting around in stripes-connect and stripes-core,
I've not been able to determine why ApiSettingsForm is passed both a
handleSubmit function and a onSubmit function and I've not been able to
determine the distinction between the two. So, I'm happy at this point
to use handleSubmit as it restores the saving functionality and passes
the tests.